# feat(java): optimize read float/double for jvm jit inline