--- a/java/fury-core/src/main/java/org/apache/fury/builder/CodecBuilder.java
+++ b/java/fury-core/src/main/java/org/apache/fury/builder/CodecBuilder.java
@@ -595,7 +595,12 @@ public abstract class CodecBuilder {
   }
 
   protected Expression unsafeGetShort(Expression base, Expression pos) {
-    return new StaticInvoke(MemoryBuffer.class, "unsafeGetShort", PRIMITIVE_SHORT_TYPE, base, pos);
+    StaticInvoke expr =
+        new StaticInvoke(Platform.class, "getShort", PRIMITIVE_SHORT_TYPE, base, pos);
+    if (!Platform.IS_LITTLE_ENDIAN) {
+      expr = new StaticInvoke(Short.class, "reverseBytes", PRIMITIVE_SHORT_TYPE, expr.inline());
+    }
+    return expr;
   }
 
   protected Expression unsafeGetInt(Expression base, Expression pos) {