--- a/java/fury-core/src/main/java/org/apache/fury/builder/CodecBuilder.java
+++ b/java/fury-core/src/main/java/org/apache/fury/builder/CodecBuilder.java
@@ -640,16 +640,6 @@ public abstract class CodecBuilder {
     return new Invoke(buffer, func, PRIMITIVE_SHORT_TYPE);
   }
 
-  protected Expression readChar(Expression buffer) {
-    String func = Platform.IS_LITTLE_ENDIAN ? "readCharOnLE" : "readCharOnBE";
-    return new Invoke(buffer, func, PRIMITIVE_CHAR_TYPE);
-  }
-
-  protected Expression readShort(Expression buffer) {
-    String func = Platform.IS_LITTLE_ENDIAN ? "readShortOnLE" : "readShortOnBE";
-    return new Invoke(buffer, func, PRIMITIVE_SHORT_TYPE);
-  }
-
   protected Expression readInt(Expression buffer) {
     String func = Platform.IS_LITTLE_ENDIAN ? "readIntOnLE" : "readIntOnBE";
     return new Invoke(buffer, func, PRIMITIVE_INT_TYPE);