--- a/java/fury-core/src/main/java/org/apache/fury/builder/CodecBuilder.java
+++ b/java/fury-core/src/main/java/org/apache/fury/builder/CodecBuilder.java
@@ -625,9 +625,9 @@ public abstract class CodecBuilder {
   protected Expression unsafeGetDouble(Expression base, Expression pos) {
     StaticInvoke expr = new StaticInvoke(Platform.class, "getLong", PRIMITIVE_LONG_TYPE, base, pos);
     if (!Platform.IS_LITTLE_ENDIAN) {
-      expr = new StaticInvoke(Long.class, "reverseBytes", PRIMITIVE_INT_TYPE, expr.inline());
+      expr = new StaticInvoke(Long.class, "reverseBytes", PRIMITIVE_LONG_TYPE, expr.inline());
     }
-    return new StaticInvoke(Double.class, "longBitsToDouble", PRIMITIVE_FLOAT_TYPE, expr.inline());
+    return new StaticInvoke(Double.class, "longBitsToDouble", PRIMITIVE_DOUBLE_TYPE, expr.inline());
   }
 
   protected Expression readChar(Expression buffer) {