… java/fury-core/src/test/java/org/apache/fury/memory/MemoryBufferTest.java
… java/fury-core/src/test/java/org/apache/fury/serializer/PrimitiveSerializersTest.java