- optimize read float jvm jit inline by separate little/big endian methods
- optimize read double jvm jit inline by separate little/big endian methods
- generate unsafe get float code online
- generate unsafe get double code online