@chaokunyang 2024-04-06
> Hi @chaokunyang , I have a suggestion. For this kind of optimization PR, we should force the submitter to provide a benchmark data in the PR, so that we can ensure that the optimization is beneficial to the codebase.
> 
> Then for the PR of the optimization category, we should force the PR title to indicate that it is an optimization PR (standardizing the Fury community). We can add a new category `opt` in [pr-lint.yml](https://github.com/apache/incubator-fury/blob/main/.github/workflows/pr-lint.yml), and I will submit a PR to do this.

How abouting add a `perf` category, as described in https://platform.uno/docs/articles/uno-development/git-conventional-commits.html.  `opt` can represent many things, like `refactor`