@chaokunyang 2024-04-06
Here is the benchmark result:
```
With This PR:
Benchmark                                  (bufferType)   (objectType)  (references)   Mode  Cnt        Score        Error  Units
UserTypeDeserializeSuite.fury_deserialize         array  Struct         false  thrpt   10  4466936.113 ± 386617.618  ops/s
UserTypeDeserializeSuite.fury_deserialize         array  Struct          true  thrpt   10  3963160.097 ± 178393.730  ops/s
UserTypeDeserializeSuite.fury_deserialize  directBuffer  Struct         false  thrpt   10  4879409.124 ±  79998.330  ops/s
UserTypeDeserializeSuite.fury_deserialize  directBuffer  Struct          true  thrpt   10  4140782.389 ±  95492.701  ops/s

Before this PR:
Benchmark                                  (bufferType)   (objectType)  (references)   Mode  Cnt        Score        Error  Units
UserTypeDeserializeSuite.fury_deserialize         array  Struct         false  thrpt   10  4135698.223 ± 344998.039  ops/s
UserTypeDeserializeSuite.fury_deserialize         array  Struct          true  thrpt   10  4062397.644 ±  94063.642  ops/s
UserTypeDeserializeSuite.fury_deserialize  directBuffer  Struct         false  thrpt   10  4311519.019 ± 251129.086  ops/s
UserTypeDeserializeSuite.fury_deserialize  directBuffer  Struct          true  thrpt   10  4056600.395 ± 414134.048  ops/s
```